add modalities-specific bids modules
