+++ python/lib/imaging_lib/bids/eeg/dataset.py
@@ -1,0 +1,6 @@
+# TODO: Complete with EEG-specific content.
+from lib.imaging_lib.bids.dataset import BIDSDataType
+
+
+class BIDSEEGDataType(BIDSDataType):
+    pass
